feat: Optimize datetime performance with dedicated converters
Major performance improvements for datetime operations:

## DateTime INTEGER Storage System
- All datetime types now use INTEGER storage for perfect microsecond precision
- DATE: INTEGER days since Unix epoch
- TIME/TIMETZ: INTEGER microseconds since midnight
- TIMESTAMP/TIMESTAMPTZ: INTEGER microseconds since Unix epoch
- INTERVAL: INTEGER microseconds duration

## Performance Optimizations
- Added dedicated type converters (indices 6, 7, 8) for date/time/timestamp types
- Implemented buffer-based formatting functions avoiding string allocations:
  - format_days_to_date_buf() for DATE
  - format_microseconds_to_time_buf() for TIME
  - format_microseconds_to_timestamp_buf() for TIMESTAMP
- Updated all hot paths to use optimized converters:
  - execute_with_cached_metadata()
  - execute_cached_query()
  - query_fast_path()

## Benchmark Results
Performance improvements achieved:
- SELECT: 21% faster (196x → 154x overhead)
- SELECT (cached): 7% faster (30x → 28x overhead)
- UPDATE: 11% faster (40x → 36x overhead)
- DELETE: 11% faster (45x → 40x overhead)
- INSERT: 6% faster (220x → 206x overhead)

## Migration System Updates
- In-memory databases now auto-migrate on startup
- Added migration v4 for datetime INTEGER storage conversion
- Updated documentation to reflect auto-migration behavior

## Other Changes
- Fixed all datetime-related tests to handle new INTEGER types
- Updated type-mapping-prd.md with comprehensive datetime documentation
- Enhanced README with performance results and datetime details

+++ src/types/type_mapper.rs
@@ -147,10 +147,10 @@ impl TypeMapper {
         mapper.pg_to_sqlite.insert("uuid".to_string(), "TEXT".to_string());
         mapper.pg_to_sqlite.insert("json".to_string(), "TEXT".to_string());
         mapper.pg_to_sqlite.insert("jsonb".to_string(), "TEXT".to_string());
-        mapper.pg_to_sqlite.insert("date".to_string(), "REAL".to_string());
-        mapper.pg_to_sqlite.insert("time".to_string(), "REAL".to_string());
-        mapper.pg_to_sqlite.insert("timestamp".to_string(), "REAL".to_string());
-        mapper.pg_to_sqlite.insert("timestamptz".to_string(), "REAL".to_string());
+        mapper.pg_to_sqlite.insert("date".to_string(), "INTEGER".to_string());
+        mapper.pg_to_sqlite.insert("time".to_string(), "INTEGER".to_string());
+        mapper.pg_to_sqlite.insert("timestamp".to_string(), "INTEGER".to_string());
+        mapper.pg_to_sqlite.insert("timestamptz".to_string(), "INTEGER".to_string());
         mapper.pg_to_sqlite.insert("numeric".to_string(), "DECIMAL".to_string());
         mapper.pg_to_sqlite.insert("decimal".to_string(), "DECIMAL".to_string());
         mapper.pg_to_sqlite.insert("bytea".to_string(), "BLOB".to_string());
@@ -160,12 +160,12 @@ impl TypeMapper {
         mapper.pg_to_sqlite.insert("bigserial".to_string(), "INTEGER".to_string());
         mapper.pg_to_sqlite.insert("character varying".to_string(), "TEXT".to_string());
         mapper.pg_to_sqlite.insert("character".to_string(), "TEXT".to_string());
-        mapper.pg_to_sqlite.insert("timestamp with time zone".to_string(), "REAL".to_string());
-        mapper.pg_to_sqlite.insert("timestamp without time zone".to_string(), "REAL".to_string());
-        mapper.pg_to_sqlite.insert("time with time zone".to_string(), "REAL".to_string());
-        mapper.pg_to_sqlite.insert("time without time zone".to_string(), "REAL".to_string());
-        mapper.pg_to_sqlite.insert("timetz".to_string(), "REAL".to_string());
-        mapper.pg_to_sqlite.insert("interval".to_string(), "REAL".to_string());
+        mapper.pg_to_sqlite.insert("timestamp with time zone".to_string(), "INTEGER".to_string());
+        mapper.pg_to_sqlite.insert("timestamp without time zone".to_string(), "INTEGER".to_string());
+        mapper.pg_to_sqlite.insert("time with time zone".to_string(), "INTEGER".to_string());
+        mapper.pg_to_sqlite.insert("time without time zone".to_string(), "INTEGER".to_string());
+        mapper.pg_to_sqlite.insert("timetz".to_string(), "INTEGER".to_string());
+        mapper.pg_to_sqlite.insert("interval".to_string(), "INTEGER".to_string());
         
         // New type mappings
         mapper.pg_to_sqlite.insert("money".to_string(), "TEXT".to_string());
@@ -492,8 +492,8 @@ mod tests {
         
         // Test multi-word types
         assert_eq!(mapper.pg_to_sqlite_for_create_table("DOUBLE PRECISION"), "DECIMAL");
-        assert_eq!(mapper.pg_to_sqlite_for_create_table("TIME WITH TIME ZONE"), "REAL");
-        assert_eq!(mapper.pg_to_sqlite_for_create_table("TIMESTAMP WITHOUT TIME ZONE"), "REAL");
+        assert_eq!(mapper.pg_to_sqlite_for_create_table("TIME WITH TIME ZONE"), "INTEGER");
+        assert_eq!(mapper.pg_to_sqlite_for_create_table("TIMESTAMP WITHOUT TIME ZONE"), "INTEGER");
         assert_eq!(mapper.pg_to_sqlite_for_create_table("BIT VARYING"), "TEXT");
     }
